requirements.txt
absl-py==0.7.1
appnope==0.1.0
astor==0.8.0
attrs==19.1.0
bleach==1.5.0
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
gast==0.2.2
google-pasta==0.1.7
grpcio==1.22.0
h5py==2.9.0
html5lib==0.9999999
imageio==2.5.0
imgaug==0.2.9
ipykernel==5.1.1
ipython==5.0.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
Jinja2==2.10.1
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kerascv==0.0.36
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.3
notebook==6.0.0
numpy==1.16.4
opencv-python==4.1.0.25
pandas==0.24.2
pandocfilters==1.4.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
prometheus-client==0.7.1
prompt-toolkit==1.0.16
protobuf==3.9.0
ptyprocess==0.6.0
Pygments==2.4.2
pyparsing==2.4.2
pyrsistent==0.15.4
python-dateutil==2.8.0
pytz==2019.1
PyWavelets==1.0.3
PyYAML==5.1.1
pyzmq==18.0.2
qtconsole==4.5.2
scikit-image==0.15.0
scipy==1.3.0
Send2Trash==1.5.0
Shapely==1.6.4.post2
simplegeneric==0.8.1
six==1.12.0
tb-nightly==1.14.0a20190603
tensorboard==1.14.0
tensorflow-gpu==2.0.0rc0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tf-estimator-nightly==1.14.0.dev2019060501
tornado==6.0.3
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.15.5
widgetsnbextension==3.5.1
wrapt==1.11.2
